parse level content
